+++ src/GLFW/Input/GlfwMouse.hpp
@@ -57,10 +57,10 @@ public:
     int getX() override;
     int getY() override;
     Point2 mouseMovement() override;
-    std::pair<double, double> getAxisFractional() override;
-    double getXFractional() override;
-    double getYFractional() override;
-    std::pair<double, double> mouseMovementFractional() override;
+    std::pair<float, float> getAxisFractional() override;
+    float getXFractional() override;
+    float getYFractional() override;
+    std::pair<float, float> mouseMovementFractional() override;
     bool mouseMoved() override;
     void warp(const Point2 &windowPosition) override;
